+++ src/main/java/com/raoulvdberge/refinedstorage/proxy/ProxyClient.java
@@ -3,6 +3,7 @@ package com.raoulvdberge.refinedstorage.proxy;
 import com.raoulvdberge.refinedstorage.RS;
 import com.raoulvdberge.refinedstorage.RSBlocks;
 import com.raoulvdberge.refinedstorage.RSItems;
+import com.raoulvdberge.refinedstorage.apiimpl.autocrafting.CraftingPattern;
 import com.raoulvdberge.refinedstorage.block.*;
 import com.raoulvdberge.refinedstorage.gui.GuiCraftingPreview;
 import com.raoulvdberge.refinedstorage.gui.grid.GuiCraftingStart;
@@ -24,6 +25,7 @@ import net.minecraft.client.renderer.VertexBuffer;
 import net.minecraft.client.renderer.block.model.ModelBakery;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
 import net.minecraft.client.renderer.block.statemap.StateMap;
+import net.minecraft.client.renderer.color.ItemColors;
 import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
 import net.minecraft.client.resources.IResourceManager;
 import net.minecraft.entity.player.EntityPlayer;
@@ -41,6 +43,7 @@ import net.minecraftforge.client.model.ModelLoader;
 import net.minecraftforge.client.model.ModelLoaderRegistry;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.FMLCommonHandler;
+import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
@@ -247,6 +250,23 @@ public class ProxyClient extends ProxyCommon {
 
             return new ModelResourceLocation("refinedstorage:controller", "direction=north,energy=" + energy);
         });
+    }
+
+    @Override
+    public void init(FMLInitializationEvent e) {
+        super.init(e);
+
+        // Register IItemColor to handle passthrough
+        ItemColors mcColors = Minecraft.getMinecraft().getItemColors();
+        mcColors.registerItemColorHandler((stack, tintIndex) -> {
+            CraftingPattern pattern = ItemPattern.getPatternFromCache(Minecraft.getMinecraft().world, stack);
+
+            if (BakedModelPattern.canDisplayPatternOutput(pattern) && mcColors.getColorFromItemstack(pattern.getOutputs().get(0), tintIndex) != -1) {
+                return mcColors.getColorFromItemstack(pattern.getOutputs().get(0), tintIndex);
+            }
+
+            return 0xFFFFFF;
+        }, RSItems.PATTERN);
     }
 
     public static void onReceiveCraftingPreviewResponse(MessageGridCraftingPreviewResponse message) {