+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.2.23
 - Fixed crash in storage cache (raoulvdberge)
+- Fixed Crafting Pattern model (pauljoda)
 
 ### 1.2.22
 - Fixed recipe for Processing Pattern Encoder not using oredictionary for the workbench (VT-14)